fixes many errors

+++ Spaceship.pde
@@ -26,13 +26,13 @@ class Spaceship extends Floater
     }
 
     public void setX(int x){myCenterX = x;}
-    public double getX(){return myCenterX;}
+    public int getX(){return (int)myCenterX;}
     public void setY(int y){myCenterY = y;}
-    public double getY(){return myCenterY;}
+    public int getY(){return (int)myCenterY;}
     public void setDirectionX(double x){myDirectionX = x;}
     public double getDirectionX(){return myDirectionX;}
     public void setDirectionY(double y){myDirectionY = y;}
     public double getDirectionY(){return myDirectionY;}
-    public void setPointDirection(double degrees){myPointDirection = degrees;}
+    public void setPointDirection(int degrees){myPointDirection = degrees;}
     public double getPointDirection(){return myPointDirection;}
 }
